re-organised lys xstate files
