+++ plugins/woocommerce-admin/client/launch-store/main-content/pages/loading.tsx
@@ -5,7 +5,7 @@ import classnames from 'classnames';
 /**
  * Internal dependencies
  */
-import type { MainContentComponentProps } from '../main-content';
+import type { MainContentComponentProps } from '../xstate';
 export const LoadingPage = ( props: MainContentComponentProps ) => {
 	return (
 		<div
