+++ plugins/woocommerce-admin/client/launch-store/main-content/pages/site-preview.tsx
@@ -5,7 +5,7 @@ import classnames from 'classnames';
 /**
  * Internal dependencies
  */
-import type { MainContentComponentProps } from '../main-content';
+import type { MainContentComponentProps } from '../xstate';
 
 export const SitePreviewPage = ( props: MainContentComponentProps ) => {
 	return (
